bsm-agentos/requirements.txt
pydantic>=2.6.0
PyYAML>=6.0.1
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.20.0
scikit-learn>=1.4.0
pytest>=8.0.0
